requirements.txt
streamlit==1.10
snowflake-connector-python